tech - rename, restructure helper functions
