+++ ui/server/helpers/mappings/map-percentage-of-cover.js
